Move Header and Main to separate components
