--- components/main/main.js
@@ -1,9 +1,0 @@
-import styles from './main.module.scss';
-
-export default function Main() {
-    return (
-        <div className={styles['main']}>
-
-        </div>
-    )
-}
