--- components/paralax/paralax.jsx
@@ -1,9 +1,0 @@
-import styles from './paralax.module.scss';
-
-export default function Paralax() {
-    return (
-        <div className={styles['paralax']}>
-
-        </div>
-    )
-}
