+++ styles/globals.scss
@@ -20,6 +20,15 @@ html {
         -apple-system, BlinkMacSystemFont,
         "Segoe UI", Roboto, Helvetica, Arial, sans-serif,
         "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
+
+    scrollbar-width: none; // Firefox
+    -ms-overflow-style: none; // IE 10+
+
+    // Blink (Chrome) & WebKit (Safari)
+    &::-webkit-scrollbar {
+        width: 0;
+        height: 0;
+    }
 }
 
 @media (prefers-color-scheme: dark) {
